+++ src/Smartstore.Web/Areas/Admin/Models/Orders/DashboardLatestOrdersModel.cs
@@ -11,7 +11,7 @@ namespace Smartstore.Admin.Models.Orders
     {
         public int CustomerId { get; set; }
         public string CustomerDisplayName { get; set; }
-        public int ProductsTotal { get; set; }
+        public decimal ProductsTotal { get; set; }
         public Money TotalAmount { get; set; }
         public string Created { get; set; }
         public OrderStatus OrderState { get; set; }
